Use BigFloat in orthogonality tests after fix to SpecialFunctions.gamma

+++ test/orthog.jl
@@ -3,8 +3,8 @@
 using GaussQuadrature
 using Printf
 
-#T = BigFloat
-T = Float64
+T = BigFloat
+#T = Float64
 println("\nFloating point type data type is ", T)
 println("\teps = ", eps(T))
